Research_Assistant/requirements.txt
flask==2.0.1
flask-cors==3.0.10
groq==0.4.0
python-dotenv==1.0.0
requests==2.28.1
beautifulsoup4==4.11.1
markdown==3.4.1
pyyaml==6.0
numpy==1.22.3
